requirements.txt
fastapi
pydantic
openai
typing
uvicorn
mysql-connector-python